change to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager first = new SalesManager(new int[]{10, 500, 64, 79,945});
+        SalesManager first = new SalesManager(new long[]{10, 500, 64, 79,945});
         first.max();
     }
 }
